Project 1 - Landing Page
Tutorial from https://www.youtube.com/watch?v=HZv8YHYUHTU&t=636s

Quite enjoyed this project; it was another one of those projects I could see being used in an actual webpage. This was quite a useful project for me to practice using the grid method without having to use a framwork such as Bootstrap; this was something that I had heard of before but I hadn't tried it yet. This project also allowed me to use Font Awesome for the first time which was something that I hadn't used before but I can understand how useful it can be 

Finally, I was able to practice using media queries to make it a responsive webpage; I had done this before once but I hadn't practiced it again in a long time so it was good to get some practice with this